PIM-6338: move filling values logic to a dedicated service
