+++ src/Pim/Bundle/CatalogBundle/DependencyInjection/PimCatalogExtension.php
@@ -57,6 +57,7 @@ class PimCatalogExtension extends Extension
         $loader->load('savers.yml');
         $loader->load('updaters.yml');
         $loader->load('validators.yml');
+        $loader->load('values_fillers.yml');
         $loader->load('versions.yml');
         $loader->load('serializers.yml');
         $loader->load('serializers_indexing.yml');
