+++ src/Pim/Bundle/CatalogBundle/Resources/config/builders.yml
@@ -10,11 +10,9 @@ services:
         arguments:
             - '@pim_catalog.repository.attribute'
             - '@pim_catalog.repository.family'
-            - '@pim_catalog.repository.currency'
             - '@pim_catalog.repository.association_type'
             - '@event_dispatcher'
             - '@pim_catalog.resolver.attribute_values'
-            - '@pim_catalog.factory.value'
             - '@pim_catalog.builder.entity_with_values'
             - {'product': '%pim_catalog.entity.product.class%', 'association': '%pim_catalog.entity.association.class%'}
 
@@ -23,11 +21,9 @@ services:
         arguments:
             - '@pim_catalog.repository.attribute'
             - '@pim_catalog.repository.family'
-            - '@pim_catalog.repository.currency'
             - '@pim_catalog.repository.association_type'
             - '@event_dispatcher'
             - '@pim_catalog.resolver.attribute_values'
-            - '@pim_catalog.factory.value'
             - '@pim_catalog.builder.entity_with_values'
             - {'product': '%pim_catalog.entity.variant_product.class%', 'association': '%pim_catalog.entity.association.class%'}
 
@@ -41,6 +37,7 @@ services:
         class: '%pim_catalog.builder.product_template.class%'
         arguments:
             - '@pim_catalog.builder.product'
+            - '@pim_catalog.values_filler.product'
             - '%pim_catalog.entity.product_template.class%'
 
     pim_catalog.builder.choices:
