+++ src/Pim/Bundle/EnrichBundle/Resources/config/connector/processors.yml
@@ -56,7 +56,7 @@ services:
             - '@pim_catalog.normalizer.standard.product'
             - '@pim_catalog.repository.channel'
             - '@pim_catalog.repository.attribute'
-            - '@pim_catalog.builder.product'
+            - '@pim_catalog.values_filler.product'
             - '@akeneo_storage_utils.doctrine.object_detacher'
             - '@pim_user.provider.user'
             - '@security.token_storage'
