+++ src/Pim/Bundle/EnrichBundle/Resources/config/controllers.yml
@@ -97,6 +97,7 @@ services:
             - '@pim_catalog.saver.product'
             - '@pim_enrich.manager.sequential_edit'
             - '@pim_catalog.builder.product'
+            - '@pim_catalog.values_filler.product'
             - '%pim_catalog.entity.category.class%'
 
     pim_enrich.controller.association:
