+++ src/Pim/Bundle/EnrichBundle/Resources/config/normalizers.yml
@@ -64,6 +64,7 @@ services:
             - '@pim_catalog.completeness.calculator'
             - '@pim_enrich.normalizer.file'
             - '@pim_catalog.builder.product'
+            - '@pim_catalog.values_filler.product'
         tags:
             - { name: pim_internal_api_serializer.normalizer }
