+++ src/Pim/Component/Catalog/spec/Builder/ProductBuilderSpec.php
@@ -3,9 +3,7 @@
 namespace spec\Pim\Component\Catalog\Builder;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Catalog\AttributeTypes;
 use Pim\Component\Catalog\Builder\EntityWithValuesBuilderInterface;
-use Pim\Component\Catalog\Factory\ValueFactory;
 use Pim\Component\Catalog\Manager\AttributeValuesResolverInterface;
 use Pim\Component\Catalog\Model\Association;
 use Pim\Component\Catalog\Model\AssociationTypeInterface;
@@ -17,7 +15,6 @@ use Pim\Component\Catalog\Model\ValueInterface;
 use Pim\Component\Catalog\ProductEvents;
 use Pim\Component\Catalog\Repository\AssociationTypeRepositoryInterface;
 use Pim\Component\Catalog\Repository\AttributeRepositoryInterface;
-use Pim\Component\Catalog\Repository\CurrencyRepositoryInterface;
 use Pim\Component\Catalog\Repository\FamilyRepositoryInterface;
 use Prophecy\Argument;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
@@ -30,12 +27,10 @@ class ProductBuilderSpec extends ObjectBehavior
     function let(
         AttributeRepositoryInterface $attributeRepository,
         FamilyRepositoryInterface $familyRepository,
-        CurrencyRepositoryInterface $currencyRepository,
         AssociationTypeRepositoryInterface $assocTypeRepository,
         EventDispatcherInterface $eventDispatcher,
         AttributeValuesResolverInterface $valuesResolver,
-        EntityWithValuesBuilderInterface $entityWithValuesBuilder,
-        ValueFactory $productValueFactory
+        EntityWithValuesBuilderInterface $entityWithValuesBuilder
     ) {
         $entityConfig = [
             'product' => self::PRODUCT_CLASS,
@@ -45,11 +40,9 @@ class ProductBuilderSpec extends ObjectBehavior
         $this->beConstructedWith(
             $attributeRepository,
             $familyRepository,
-            $currencyRepository,
             $assocTypeRepository,
             $eventDispatcher,
             $valuesResolver,
-            $productValueFactory,
             $entityWithValuesBuilder,
             $entityConfig
         );
@@ -90,94 +83,6 @@ class ProductBuilderSpec extends ObjectBehavior
         $product->shouldReturnAnInstanceOf(self::PRODUCT_CLASS);
     }
 
-    function it_adds_missing_product_values_from_family_on_new_product(
-        $valuesResolver,
-        $entityWithValuesBuilder,
-        FamilyInterface $family,
-        ProductInterface $product,
-        AttributeInterface $sku,
-        AttributeInterface $name,
-        AttributeInterface $desc,
-        ValueInterface $skuValue
-    ) {
-        $sku->getCode()->willReturn('sku');
-        $sku->getType()->willReturn('pim_catalog_identifier');
-        $sku->isLocalizable()->willReturn(false);
-        $sku->isScopable()->willReturn(false);
-
-        $name->getCode()->willReturn('name');
-        $name->getType()->willReturn('pim_catalog_text');
-        $name->isLocalizable()->willReturn(true);
-        $name->isScopable()->willReturn(false);
-
-        $desc->getCode()->willReturn('description');
-        $desc->getType()->willReturn('pim_catalog_text');
-        $desc->isLocalizable()->willReturn(true);
-        $desc->isScopable()->willReturn(true);
-
-        // get expected attributes
-        $product->getAttributes()->willReturn([$sku]);
-        $family->getAttributes()->willReturn([$sku, $name, $desc]);
-        $product->getFamily()->willReturn($family);
-
-        // get eligible values
-        $valuesResolver->resolveEligibleValues(['sku' => $sku, 'name' => $name, 'description' => $desc], null, null)
-            ->willReturn([
-                [
-                    'attribute' => 'sku',
-                    'type' => 'pim_catalog_identifier',
-                    'locale' => null,
-                    'scope' => null
-                ],
-                [
-                    'attribute' => 'name',
-                    'type' => 'pim_catalog_text',
-                    'locale' => 'fr_FR',
-                    'scope' => null
-                ],
-                [
-                    'attribute' => 'name',
-                    'type' => 'pim_catalog_text',
-                    'locale' => 'en_US',
-                    'scope' => null
-                ],
-                [
-                    'attribute' => 'description',
-                    'type' => 'pim_catalog_text',
-                    'locale' => 'en_US',
-                    'scope' => 'ecommerce'
-                ],
-                [
-                    'attribute' => 'description',
-                    'type' => 'pim_catalog_text',
-                    'locale' => 'fr_FR',
-                    'scope' => 'ecommerce'
-                ],
-                [
-                    'attribute' => 'description',
-                    'type' => 'pim_catalog_text',
-                    'locale' => 'en_US',
-                    'scope' => 'print'
-                ],
-                [
-                    'attribute' => 'description',
-                    'type' => 'pim_catalog_text',
-                    'locale' => 'fr_FR',
-                    'scope' => 'print'
-                ]
-            ]);
-
-        // get existing values
-        $skuValue->getAttribute()->willReturn($sku);
-        $skuValue->getLocale()->willReturn(null);
-        $skuValue->getScope()->willReturn(null);
-        $product->getValues()->willReturn([$skuValue]);
-
-        $entityWithValuesBuilder->addOrReplaceValue(Argument::cetera())->shouldBeCalledTimes(6);
-
-        $this->addMissingProductValues($product);
-    }
-
     function it_adds_missing_product_associations(
         $assocTypeRepository,
         ProductInterface $productOne,
@@ -191,95 +96,5 @@ class ProductBuilderSpec extends ObjectBehavior
         $assocTypeRepository->findMissingAssociationTypes($productTwo)->willReturn([]);
         $productTwo->addAssociation(Argument::any())->shouldNotBeCalled();
         $this->addMissingAssociations($productTwo);
-    }
-
-    function it_adds_an_empty_product_value(
-        $productValueFactory,
-        ProductInterface $product,
-        AttributeInterface $size,
-        AttributeInterface $color,
-        ValueInterface $sizeValue,
-        ValueInterface $colorValue
-    ) {
-        $size->getCode()->willReturn('size');
-        $size->getType()->willReturn(AttributeTypes::OPTION_SIMPLE_SELECT);
-        $size->isLocalizable()->willReturn(false);
-        $size->isScopable()->willReturn(false);
-
-        $color->getCode()->willReturn('color');
-        $color->getType()->willReturn(AttributeTypes::OPTION_SIMPLE_SELECT);
-        $color->isLocalizable()->willReturn(true);
-        $color->isScopable()->willReturn(true);
-
-        $product->getValue('size', null, null)->willReturn($sizeValue);
-        $product->getValue('color', 'en_US', 'ecommerce')->willReturn($colorValue);
-
-        $product->removeValue($sizeValue)->willReturn($product);
-        $product->removeValue($colorValue)->willReturn($product);
-
-        $productValueFactory->create($size, null, null, null)->willReturn($sizeValue);
-        $productValueFactory->create($color, 'ecommerce', 'en_US', null)->willReturn($colorValue);
-
-        $product->addValue($sizeValue)->willReturn($product);
-        $product->addValue($colorValue)->willReturn($product);
-
-        $this->addOrReplaceValue($product, $size, null, null, null);
-        $this->addOrReplaceValue($product, $color, 'en_US', 'ecommerce', null);
-    }
-
-    function it_adds_a_non_empty_product_value(
-        $productValueFactory,
-        ProductInterface $product,
-        AttributeInterface $size,
-        AttributeInterface $color,
-        ValueInterface $sizeValue,
-        ValueInterface $colorValue
-    ) {
-        $size->getCode()->willReturn('size');
-        $size->getType()->willReturn(AttributeTypes::OPTION_SIMPLE_SELECT);
-        $size->isLocalizable()->willReturn(false);
-        $size->isScopable()->willReturn(false);
-
-        $color->getCode()->willReturn('color');
-        $color->getType()->willReturn(AttributeTypes::OPTION_SIMPLE_SELECT);
-        $color->isLocalizable()->willReturn(true);
-        $color->isScopable()->willReturn(true);
-
-        $product->getValue('size', null, null)->willReturn($sizeValue);
-        $product->getValue('color', 'en_US', 'ecommerce')->willReturn($colorValue);
-
-        $product->removeValue($sizeValue)->willReturn($product);
-        $product->removeValue($colorValue)->willReturn($product);
-
-        $productValueFactory->create($size, null, null, null)->willReturn($sizeValue);
-        $productValueFactory->create($color, 'ecommerce', 'en_US', 'red')->willReturn($colorValue);
-
-        $product->addValue($sizeValue)->willReturn($product);
-        $product->addValue($colorValue)->willReturn($product);
-
-        $this->addOrReplaceValue($product, $size, null, null, null);
-        $this->addOrReplaceValue($product, $color, 'en_US', 'ecommerce', 'red');
-    }
-
-    function it_adds_a_product_value_if_there_was_not_a_previous_one(
-        $productValueFactory,
-        ProductInterface $product,
-        AttributeInterface $label,
-        ValueInterface $value
-    ) {
-        $label->getCode()->willReturn('label');
-        $label->getType()->willReturn(AttributeTypes::TEXT);
-        $label->isLocalizable()->willReturn(false);
-        $label->isScopable()->willReturn(false);
-
-        $product->getValue('label', null, null)->willReturn(null);
-
-        $product->removeValue(Argument::any())->shouldNotBeCalled();
-
-        $productValueFactory->create($label, null, null, 'foobar')->willReturn($value);
-
-        $product->addValue($value)->willReturn($product);
-
-        $this->addOrReplaceValue($product, $label, null, null, 'foobar');
     }
 }
