+++ src/Pim/Component/Catalog/spec/ValuesFiller/ProductValuesFillerSpec.php
@@ -1,0 +1,113 @@
+<?php
+
+namespace spec\Pim\Component\Catalog\ValuesFiller;
+
+use PhpSpec\ObjectBehavior;
+use Pim\Component\Catalog\Builder\EntityWithValuesBuilderInterface;
+use Pim\Component\Catalog\Builder\ProductBuilderInterface;
+use Pim\Component\Catalog\Manager\AttributeValuesResolverInterface;
+use Pim\Component\Catalog\Model\AttributeInterface;
+use Pim\Component\Catalog\Model\FamilyInterface;
+use Pim\Component\Catalog\Model\ProductInterface;
+use Pim\Component\Catalog\Model\ValueInterface;
+use Pim\Component\Catalog\Repository\CurrencyRepositoryInterface;
+use Prophecy\Argument;
+
+class ProductValuesFillerSpec extends ObjectBehavior
+{
+    function let(
+        EntityWithValuesBuilderInterface $entityWithValuesBuilder,
+        AttributeValuesResolverInterface $valuesResolver,
+        CurrencyRepositoryInterface $currencyRepository
+    ) {
+        $this->beConstructedWith($entityWithValuesBuilder, $valuesResolver, $currencyRepository);
+    }
+
+    function it_fills_missing_product_values_from_family_on_new_product(
+        $valuesResolver,
+        $entityWithValuesBuilder,
+        FamilyInterface $family,
+        ProductInterface $product,
+        AttributeInterface $sku,
+        AttributeInterface $name,
+        AttributeInterface $desc,
+        ValueInterface $skuValue
+    ) {
+        $sku->getCode()->willReturn('sku');
+        $sku->getType()->willReturn('pim_catalog_identifier');
+        $sku->isLocalizable()->willReturn(false);
+        $sku->isScopable()->willReturn(false);
+
+        $name->getCode()->willReturn('name');
+        $name->getType()->willReturn('pim_catalog_text');
+        $name->isLocalizable()->willReturn(true);
+        $name->isScopable()->willReturn(false);
+
+        $desc->getCode()->willReturn('description');
+        $desc->getType()->willReturn('pim_catalog_text');
+        $desc->isLocalizable()->willReturn(true);
+        $desc->isScopable()->willReturn(true);
+
+        // get expected attributes
+        $product->getAttributes()->willReturn([$sku]);
+        $family->getAttributes()->willReturn([$sku, $name, $desc]);
+        $product->getFamily()->willReturn($family);
+
+        // get eligible values
+        $valuesResolver->resolveEligibleValues(['sku' => $sku, 'name' => $name, 'description' => $desc], null, null)
+            ->willReturn([
+                [
+                    'attribute' => 'sku',
+                    'type' => 'pim_catalog_identifier',
+                    'locale' => null,
+                    'scope' => null
+                ],
+                [
+                    'attribute' => 'name',
+                    'type' => 'pim_catalog_text',
+                    'locale' => 'fr_FR',
+                    'scope' => null
+                ],
+                [
+                    'attribute' => 'name',
+                    'type' => 'pim_catalog_text',
+                    'locale' => 'en_US',
+                    'scope' => null
+                ],
+                [
+                    'attribute' => 'description',
+                    'type' => 'pim_catalog_text',
+                    'locale' => 'en_US',
+                    'scope' => 'ecommerce'
+                ],
+                [
+                    'attribute' => 'description',
+                    'type' => 'pim_catalog_text',
+                    'locale' => 'fr_FR',
+                    'scope' => 'ecommerce'
+                ],
+                [
+                    'attribute' => 'description',
+                    'type' => 'pim_catalog_text',
+                    'locale' => 'en_US',
+                    'scope' => 'print'
+                ],
+                [
+                    'attribute' => 'description',
+                    'type' => 'pim_catalog_text',
+                    'locale' => 'fr_FR',
+                    'scope' => 'print'
+                ]
+            ]);
+
+        // get existing values
+        $skuValue->getAttribute()->willReturn($sku);
+        $skuValue->getLocale()->willReturn(null);
+        $skuValue->getScope()->willReturn(null);
+        $product->getValues()->willReturn([$skuValue]);
+
+        $entityWithValuesBuilder->addOrReplaceValue(Argument::cetera())->shouldBeCalledTimes(6);
+
+        $this->fillMissingValues($product);
+    }
+}
